requirements.txt
asgiref==3.6.0
Django==4.2.1
django-filebrowser-no-grappelli
django-ckeditor==6.5.1
django-js-asset==2.0.0
django-tinymce
easy-thumbnails==2.8.5
jsmin==3.0.1
Pillow==9.5.0
rptree==0.1.1
six==1.16.0
sqlparse==0.4.4
django-storages
boto3
